% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_vector.R
\name{get_results_vector}
\alias{get_results_vector}
\title{Get Results Vector}
\usage{
get_results_vector(
  hypothesis,
  differences,
  diff_method = "wilcoxon",
  grp_a = NULL,
  grp_b = NULL,
  phi_0 = 0.5
)
}
\arguments{
\item{hypothesis}{A string or a vector of strings where the string or all elements of the vector are in the set of {'decrease', 'increase', 'different'}. If it’s a string, it will be converted to a vector of the same length as \code{differences} with all elements being the valid string.}

\item{differences}{A numeric vector representing the differences between two groups.}

\item{diff_method}{A string specifying the method to use for testing 'different' hypotheses. Valid options are 'wilcoxon' or 't'. Defaults to 'wilcoxon'.}

\item{grp_a}{A data frame representing the first group for testing 'different' hypotheses. This argument is only needed if 'different' is part of the hypothesis.}

\item{grp_b}{A data frame representing the second group for testing 'different' hypotheses. This argument is only needed if 'different' is part of the hypothesis.}

\item{phi_0}{A numeric value on the interval (0, 1) representing the decision rule threshold for the p-value. Defaults to 0.5.}
}
\value{
A numeric vector of 0s and 1s. If the hypothesis was incorrect, a 0 is returned. If the hypothesis was correct, a 1 is returned.
}
\description{
This function receives user input on the hypothesis of the experiment results and informs the user if the hypotheses were correct. It can handle hypotheses of 'increase', 'decrease', or 'different', and performs appropriate statistical tests.
}
\details{
The function checks if the input hypotheses are valid, and performs the necessary statistical tests to determine if the hypotheses were correct. It handles 'increase' and 'decrease' hypotheses by comparing differences, and 'different' hypotheses by performing either a Wilcoxon signed-rank test or a paired t-test.
}
\examples{
data("group_data_example")
data("group_cog_data")
data("pre_post_data_example")
data("pre_post_fit")

# Example 1: Using group_data_example with 'increase' hypothesis
grp_1_data <- subset(group_data_example, group == 'placebo', select = c('v1', 'v2'))
grp_2_data <- subset(group_data_example, group == 'drug', select = c('v1', 'v2'))
differences <- create_difference_vector(grp_1_data, grp_2_data, location='mean')
get_results_vector(hypothesis = 'increase', differences = differences)

# Example 2: Using pre_post_data_example with 'decrease' hypothesis
pre_data <- subset(pre_post_data_example, time == 0, select = c('v1', 'v2', 'v3'))
post_data <- subset(pre_post_data_example, time == 12, select = c('v1', 'v2', 'v3'))
differences <- create_difference_vector(pre_data, post_data, location='median')
get_results_vector(hypothesis = 'decrease', differences = differences)

# Example 3: Using group_cog_data with mixed 'increase' and 'decrease' hypothesis
control_data <- subset(group_cog_data, group.factor == 'Control', select = c('blind_moca_uncorrected', 'craft_verbatim'))
eskd_data <- subset(group_cog_data, group.factor == 'ESKD', select = c('blind_moca_uncorrected', 'craft_verbatim'))
differences <- create_difference_vector(control_data, eskd_data, location='mean')
get_results_vector(hypothesis = c('increase', 'decrease'), differences = differences)

# Example 4: Using pre_post_fit with 'different' hypothesis and Wilcoxon test
pre_fit <- subset(pre_post_fit, Time == 0, select = c('Flex_right', 'Flex_left'))
post_fit <- subset(pre_post_fit, Time == 1, select = c('Flex_right', 'Flex_left'))
differences <- create_difference_vector(pre_fit, post_fit, location='mean')
get_results_vector(hypothesis = 'different', differences = differences, diff_method = 'wilcoxon', grp_a = pre_fit, grp_b = post_fit)


# Test with 'increase' and 'decrease' hypotheses
get_results_vector(hypothesis = c('increase', 'decrease', 'increase'), differences = differences)

# Test with 'different' hypothesis
get_results_vector(hypothesis = c('increase', 'different', 'decrease'), differences = differences, grp_a = grp_a, grp_b = grp_b, phi_0 = 0.05)

}
