% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_weights.R
\name{pred_weights}
\alias{pred_weights}
\title{Calculate Predictive Weights}
\usage{
pred_weights(
  dataset,
  vars,
  gtvar,
  type = "group",
  id = NULL,
  pre = NULL,
  post = NULL,
  corr_method = "pearson"
)
}
\arguments{
\item{dataset}{A data frame containing the dataset to be analyzed.}

\item{vars}{A vector of strings specifying the names of the variables to be used in the correlation analysis.}

\item{gtvar}{The name of the categorical variable used to identify groups in 'prepost' type analysis.}

\item{type}{The type of analysis. Valid values are 'group' for group-based correlation analysis or 'prepost' for pre-post analysis. Defaults to 'group'.}

\item{id}{The variable in the dataset that uniquely identifies subjects in a 'prepost' analysis. This should not be \code{NULL} if \code{type} is 'prepost'.}

\item{pre}{Specifies the baseline time point for 'prepost' analysis.}

\item{post}{Specifies the follow-up time point for 'prepost' analysis.}

\item{corr_method}{The method of correlation. Valid options are 'pearson', 'kendall', or 'spearman'. Defaults to 'pearson'.}
}
\value{
A numeric vector of predictive weights for each variable analyzed.
}
\description{
This function calculates predictive weights by computing the inverse square sum of a correlation matrix derived from the specified variables. In 'group' analysis, it directly uses the variables for correlation. In 'prepost' analysis, it calculates the difference between two time points before correlation.
}
\details{
This function performs error handling to ensure appropriate input values and types. It calculates the correlation matrix for the specified variables and then computes the predictive weights as the inverse square sum of the correlation matrix.
}
\examples{
# Load example data
data("group_data_example")
data("pre_post_data_example")

# Example usage for group-based data
weights_group <- pred_weights(
  dataset = group_data_example,
  vars = c("v1", "v2", "v3"),
  gtvar = "group",
  type = "group"
)
print(weights_group)

# Example usage for pre-post data
weights_prepost <- pred_weights(
  dataset = pre_post_data_example,
  vars = c("v1", "v2", "v3"),
  gtvar = "time",
  type = "prepost",
  id = "ID",
  pre = 0,
  post = 12
)
print(weights_prepost)

}
